requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==3.7.1
appdirs==1.4.4
array-record==0.4.1
fastapi==0.103.1
fastapi_cors==0.0.6
flatbuffers==23.5.26
fonttools==4.42.0
frozenlist==1.4.0
pydantic==1.10.7
uvicorn
gunicorn
python-dateutil